agents_api/requirements.txt
fastapi
uvicorn[standard]
openai>=1.30.0
python-dotenv
requests
pydantic>=2.0.0
